+++ test/internal/DefenderDeploy.t.sol
@@ -14,21 +14,12 @@ import {NoLicense} from "../contracts/NoLicense.sol";
 import {Unlicensed} from "../contracts/Unlicensed.sol";
 import {MyContractName} from "../contracts/MyContractFile.sol";
 
+import {StringHelper} from "./StringHelper.sol";
+
 /**
  * @dev Tests the DefenderDeploy internal library.
  */
 contract DefenderDeployTest is Test {
-    function _toString(string[] memory arr) private pure returns (string memory) {
-        string memory result;
-        for (uint i = 0; i < arr.length; i++) {
-            result = string.concat(result, arr[i]);
-            if (i < arr.length - 1) {
-                result = string.concat(result, " ");
-            }
-        }
-        return result;
-    }
-
     function testBuildDeployCommand() public {
         ContractInfo memory contractInfo = Utils.getContractInfo("MyContractFile.sol:MyContractName", "out");
         string memory buildInfoFile = Utils.getBuildInfoFile(
@@ -38,7 +29,7 @@ contract DefenderDeployTest is Test {
         );
 
         DefenderOptions memory opts;
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, "", opts)
         );
 
@@ -65,7 +56,7 @@ contract DefenderDeployTest is Test {
         bytes memory constructorData = abi.encode(123);
 
         DefenderOptions memory opts;
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, constructorData, opts)
         );
 
@@ -100,8 +91,9 @@ contract DefenderDeployTest is Test {
         opts.txOverrides.gasPrice = 1 gwei;
         opts.txOverrides.maxFeePerGas = 2 gwei;
         opts.txOverrides.maxPriorityFeePerGas = 0.5 gwei;
+        opts.metadata = '{ "commitHash": "4ae3e0d", "tag": "v1.0.0", "anyOtherField": "anyValue" }';
 
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, constructorData, opts)
         );
 
@@ -112,7 +104,8 @@ contract DefenderDeployTest is Test {
                 Versions.DEFENDER_DEPLOY_CLIENT_CLI,
                 " deploy --contractName WithConstructor --contractPath test/contracts/WithConstructor.sol --chainId 31337 --buildInfoFile ",
                 buildInfoFile,
-                ' --constructorBytecode 0x000000000000000000000000000000000000000000000000000000000000007b --licenseType "My License Type" --relayerId my-relayer-id --salt 0xabc0000000000000000000000000000000000000000000000000000000000123 --gasLimit 100000 --gasPrice 1000000000 --maxFeePerGas 2000000000 --maxPriorityFeePerGas 500000000'
+                ' --constructorBytecode 0x000000000000000000000000000000000000000000000000000000000000007b --licenseType "My License Type" --relayerId my-relayer-id --salt 0xabc0000000000000000000000000000000000000000000000000000000000123 --gasLimit 100000 --gasPrice 1000000000 --maxFeePerGas 2000000000 --maxPriorityFeePerGas 500000000',
+                ' --metadata "{ \\"commitHash\\": \\"4ae3e0d\\", \\"tag\\": \\"v1.0.0\\", \\"anyOtherField\\": \\"anyValue\\" }"'
             )
         );
     }
@@ -130,7 +123,7 @@ contract DefenderDeployTest is Test {
         DefenderOptions memory opts;
         opts.skipVerifySourceCode = true;
 
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, constructorData, opts)
         );
 
@@ -159,7 +152,7 @@ contract DefenderDeployTest is Test {
         DefenderOptions memory opts;
         opts.skipLicenseType = true;
 
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, constructorData, opts)
         );
 
@@ -252,7 +245,7 @@ contract DefenderDeployTest is Test {
         );
 
         DefenderOptions memory opts;
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, "", opts)
         );
 
@@ -276,7 +269,7 @@ contract DefenderDeployTest is Test {
         );
 
         DefenderOptions memory opts;
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildDeployCommand(contractInfo, buildInfoFile, "", opts)
         );
 
@@ -296,7 +289,7 @@ contract DefenderDeployTest is Test {
         ContractInfo memory contractInfo = Utils.getContractInfo("MyContractFile.sol:MyContractName", "out");
 
         Options memory opts;
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildProposeUpgradeCommand(
                 address(0x1230000000000000000000000000000000000456),
                 address(0),
@@ -337,7 +330,7 @@ contract DefenderDeployTest is Test {
     }
 
     function testBuildGetApprovalProcessCommand() public view {
-        string memory commandString = _toString(
+        string memory commandString = StringHelper.join(
             DefenderDeploy.buildGetApprovalProcessCommand("getDeployApprovalProcess")
         );