+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.3.6 (2024-09-24)
+
+- Add `referenceBuildInfoDir` and `exclude` options. ([#74](https://github.com/OpenZeppelin/openzeppelin-foundry-upgrades/pull/74))
+
+## 0.3.5 (2024-09-20)
+
+### Potentially breaking changes
+- Adds a check to ensure `initialOwner` for `deployTransparentProxy` is not a ProxyAdmin contract. ([#76](https://github.com/OpenZeppelin/openzeppelin-foundry-upgrades/pull/76))
+
+## 0.3.4 (2024-09-16)
+
+- Defender: Add `metadata` option. ([#75](https://github.com/OpenZeppelin/openzeppelin-foundry-upgrades/pull/75))
+
 ## 0.3.3 (2024-08-26)
 
 - Defender: Fix error in `proposeUpgrade` when project path has a space. ([#71](https://github.com/OpenZeppelin/openzeppelin-foundry-upgrades/pull/71))